requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Data processing
requests==2.31.0
python-geohash==0.8.5
geopy==2.4.1
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# API clients
openweathermap==0.0.3
tweepy==4.14.0

# Notification services
firebase-admin==6.3.0 